requirements.txt
flask

# use the compatible versions with
# your OpenStack environment
python-ceilometerclient==2.4.0
python-novaclient==3.3.2

# report generator requirements
cffi
jinja2
json2html
weasyprint
xlrd
